Invisible Collision Tiles and MainMenu Room
The invisible collision tiles are created by making a 32 x 32 completely empty sprite. An object is then created using this empty sprite and then placed in the room. The object is set as solid and NOT visible so when the game loads, it is not there and will act as a floor where the tileset normally cannot.